you can add person

+++ Template/app/src/main/java/com/example/denis/kotlinmvp/model/database/repository/person/Person.kt
@@ -3,4 +3,4 @@ package com.example.denis.kotlinmvp.model.database.repository.person
 /**
  * Created by denis on 31/01/2018.
  */
-data class Person (val id: Int, val name: String, val age: Int)
+data class Person (val id: Long, val name: String, val age: Int)
